refactor/connect: extract some code to separate module
Makes it easier to read/modify code, etc.
